refactor: add possibility wrap command in the another command
