+++ plugin.d.ts
@@ -1,0 +1,11 @@
+/// <reference types="cypress" />
+
+/**
+ * Installs cy-verify-downloads
+ * @example isFileExist(on, config);
+ * @param on Cypress plugin events
+ * @param config Cypress plugin config options
+ * @returns Cypress plugin config options
+*/
+declare function _exports(on: Cypress.PluginEvents, config: Cypress.PluginConfigOptions): Cypress.PluginConfigOptions
+export = _exports;
