+++ site/index.html
@@ -7,5 +7,7 @@
        <a data-cy="small" href="./assets/small-file-10MB.zip" download>Download small file</a>
        <br/>
        <a data-cy="large" href="./assets/large-file-100MB.zip" download>Download large file</a>
+       <br/>
+       <a data-cy="pdf" href="./assets/dummy.pdf" download>Download pdf file</a>
    </body>
 </html>
